4597 - [Breakpoints] Extract Context menu
