+++ src/components/SecondaryPanes/Breakpoints.js
@@ -23,10 +23,10 @@ import { makeLocationId } from "../../utils/breakpoint";
 import { endTruncateStr } from "../../utils/utils";
 import { getFilename } from "../../utils/source";
 import { isInterrupted } from "../../utils/pause";
-import { showMenu, buildMenu } from "devtools-contextmenu";
 import CloseButton from "../shared/Button/Close";
 import "./Breakpoints.css";
 import { sortBy } from "lodash";
+import showContextMenu from "./BreakpointsContextMenu";
 
 import type { Breakpoint, Location } from "../../types";
 
@@ -104,236 +104,6 @@ class Breakpoints extends PureComponent<Props> {
     }
   }
 
-  showContextMenu(e, breakpoint) {
-    const {
-      removeBreakpoint,
-      removeBreakpoints,
-      removeAllBreakpoints,
-      toggleBreakpoints,
-      toggleAllBreakpoints,
-      toggleDisabledBreakpoint,
-      setBreakpointCondition,
-      openConditionalPanel,
-      breakpoints
-    } = this.props;
-
-    e.preventDefault();
-
-    const deleteSelfLabel = L10N.getStr("breakpointMenuItem.deleteSelf2.label");
-    const deleteAllLabel = L10N.getStr("breakpointMenuItem.deleteAll2.label");
-    const deleteOthersLabel = L10N.getStr(
-      "breakpointMenuItem.deleteOthers2.label"
-    );
-    const enableSelfLabel = L10N.getStr("breakpointMenuItem.enableSelf2.label");
-    const enableAllLabel = L10N.getStr("breakpointMenuItem.enableAll2.label");
-    const enableOthersLabel = L10N.getStr(
-      "breakpointMenuItem.enableOthers2.label"
-    );
-    const disableSelfLabel = L10N.getStr(
-      "breakpointMenuItem.disableSelf2.label"
-    );
-    const disableAllLabel = L10N.getStr("breakpointMenuItem.disableAll2.label");
-    const disableOthersLabel = L10N.getStr(
-      "breakpointMenuItem.disableOthers2.label"
-    );
-    const removeConditionLabel = L10N.getStr(
-      "breakpointMenuItem.removeCondition2.label"
-    );
-    const addConditionLabel = L10N.getStr(
-      "breakpointMenuItem.addCondition2.label"
-    );
-    const editConditionLabel = L10N.getStr(
-      "breakpointMenuItem.editCondition2.label"
-    );
-
-    const deleteSelfKey = L10N.getStr(
-      "breakpointMenuItem.deleteSelf2.accesskey"
-    );
-    const deleteAllKey = L10N.getStr("breakpointMenuItem.deleteAll2.accesskey");
-    const deleteOthersKey = L10N.getStr(
-      "breakpointMenuItem.deleteOthers2.accesskey"
-    );
-    const enableSelfKey = L10N.getStr(
-      "breakpointMenuItem.enableSelf2.accesskey"
-    );
-    const enableAllKey = L10N.getStr("breakpointMenuItem.enableAll2.accesskey");
-    const enableOthersKey = L10N.getStr(
-      "breakpointMenuItem.enableOthers2.accesskey"
-    );
-    const disableSelfKey = L10N.getStr(
-      "breakpointMenuItem.disableSelf2.accesskey"
-    );
-    const disableAllKey = L10N.getStr(
-      "breakpointMenuItem.disableAll2.accesskey"
-    );
-    const disableOthersKey = L10N.getStr(
-      "breakpointMenuItem.disableOthers2.accesskey"
-    );
-    const removeConditionKey = L10N.getStr(
-      "breakpointMenuItem.removeCondition2.accesskey"
-    );
-    const editConditionKey = L10N.getStr(
-      "breakpointMenuItem.editCondition2.accesskey"
-    );
-    const addConditionKey = L10N.getStr(
-      "breakpointMenuItem.addCondition2.accesskey"
-    );
-
-    const otherBreakpoints = breakpoints.filter(b => b !== breakpoint);
-    const enabledBreakpoints = breakpoints.filter(b => !b.disabled);
-    const disabledBreakpoints = breakpoints.filter(b => b.disabled);
-    const otherEnabledBreakpoints = breakpoints.filter(
-      b => !b.disabled && b !== breakpoint
-    );
-    const otherDisabledBreakpoints = breakpoints.filter(
-      b => b.disabled && b !== breakpoint
-    );
-
-    const deleteSelf = {
-      id: "node-menu-delete-self",
-      label: deleteSelfLabel,
-      accesskey: deleteSelfKey,
-      disabled: false,
-      click: () => removeBreakpoint(breakpoint.location)
-    };
-
-    const deleteAll = {
-      id: "node-menu-delete-all",
-      label: deleteAllLabel,
-      accesskey: deleteAllKey,
-      disabled: false,
-      click: () => removeAllBreakpoints()
-    };
-
-    const deleteOthers = {
-      id: "node-menu-delete-other",
-      label: deleteOthersLabel,
-      accesskey: deleteOthersKey,
-      disabled: false,
-      click: () => removeBreakpoints(otherBreakpoints)
-    };
-
-    const enableSelf = {
-      id: "node-menu-enable-self",
-      label: enableSelfLabel,
-      accesskey: enableSelfKey,
-      disabled: false,
-      click: () => toggleDisabledBreakpoint(breakpoint.location.line)
-    };
-
-    const enableAll = {
-      id: "node-menu-enable-all",
-      label: enableAllLabel,
-      accesskey: enableAllKey,
-      disabled: false,
-      click: () => toggleAllBreakpoints(false)
-    };
-
-    const enableOthers = {
-      id: "node-menu-enable-others",
-      label: enableOthersLabel,
-      accesskey: enableOthersKey,
-      disabled: false,
-      click: () => toggleBreakpoints(false, otherDisabledBreakpoints)
-    };
-
-    const disableSelf = {
-      id: "node-menu-disable-self",
-      label: disableSelfLabel,
-      accesskey: disableSelfKey,
-      disabled: false,
-      click: () => toggleDisabledBreakpoint(breakpoint.location.line)
-    };
-
-    const disableAll = {
-      id: "node-menu-disable-all",
-      label: disableAllLabel,
-      accesskey: disableAllKey,
-      disabled: false,
-      click: () => toggleAllBreakpoints(true)
-    };
-
-    const disableOthers = {
-      id: "node-menu-disable-others",
-      label: disableOthersLabel,
-      accesskey: disableOthersKey,
-      click: () => toggleBreakpoints(true, otherEnabledBreakpoints)
-    };
-
-    const removeCondition = {
-      id: "node-menu-remove-condition",
-      label: removeConditionLabel,
-      accesskey: removeConditionKey,
-      disabled: false,
-      click: () => setBreakpointCondition(breakpoint.location)
-    };
-
-    const addCondition = {
-      id: "node-menu-add-condition",
-      label: addConditionLabel,
-      accesskey: addConditionKey,
-      click: () => {
-        this.selectBreakpoint(breakpoint);
-        openConditionalPanel(breakpoint.location.line);
-      }
-    };
-
-    const editCondition = {
-      id: "node-menu-edit-condition",
-      label: editConditionLabel,
-      accesskey: editConditionKey,
-      click: () => {
-        this.selectBreakpoint(breakpoint);
-        openConditionalPanel(breakpoint.location.line);
-      }
-    };
-
-    const hideEnableSelf = !breakpoint.disabled;
-    const hideEnableAll = disabledBreakpoints.size === 0;
-    const hideEnableOthers = otherDisabledBreakpoints.size === 0;
-    const hideDisableAll = enabledBreakpoints.size === 0;
-    const hideDisableOthers = otherEnabledBreakpoints.size === 0;
-    const hideDisableSelf = breakpoint.disabled;
-
-    const items = [
-      { item: enableSelf, hidden: () => hideEnableSelf },
-      { item: enableAll, hidden: () => hideEnableAll },
-      { item: enableOthers, hidden: () => hideEnableOthers },
-      {
-        item: { type: "separator" },
-        hidden: () => hideEnableSelf && hideEnableAll && hideEnableOthers
-      },
-      { item: deleteSelf },
-      { item: deleteAll },
-      { item: deleteOthers, hidden: () => breakpoints.size === 1 },
-      {
-        item: { type: "separator" },
-        hidden: () => hideDisableSelf && hideDisableAll && hideDisableOthers
-      },
-
-      { item: disableSelf, hidden: () => hideDisableSelf },
-      { item: disableAll, hidden: () => hideDisableAll },
-      { item: disableOthers, hidden: () => hideDisableOthers },
-      {
-        item: { type: "separator" }
-      },
-      {
-        item: addCondition,
-        hidden: () => breakpoint.condition
-      },
-      {
-        item: editCondition,
-        hidden: () => !breakpoint.condition
-      },
-      {
-        item: removeCondition,
-        hidden: () => !breakpoint.condition
-      }
-    ];
-
-    showMenu(e, buildMenu(items));
-  }
-
   selectBreakpoint(breakpoint) {
     this.props.selectLocation(breakpoint.location);
   }
@@ -367,7 +137,9 @@ class Breakpoints extends PureComponent<Props> {
         })}
         key={locationId}
         onClick={() => this.selectBreakpoint(breakpoint)}
-        onContextMenu={e => this.showContextMenu(e, breakpoint)}
+        onContextMenu={e =>
+          showContextMenu({ ...this.props, breakpoint, contextMenuEvent: e })
+        }
       >
         <input
           type="checkbox"
